Move tokenizer module one level up
